requirements.txt
numpy==1.24.3
opencv_contrib_python==4.8.0.74
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.0.0
Pillow==10.3.0
tensorflow==2.15.1
tensorflow_intel==2.15.1
